[BE] fix : /account/totalSaving API 기능 개선
- 전체 사용자가 아닌 현 사용자의 현재까지 saving이다.

+++ BE/src/main/java/com/codesquad/rocket/web/dto/response/account/TotalSavingResponseDto.java
@@ -14,6 +14,6 @@ import lombok.ToString;
 @ToString
 public class TotalSavingResponseDto {
 
-    private Double totalPlates;
+    private Integer totalPlates;
     private Double totalSaving;
 }
